+++ HPP/heap_card.hpp
@@ -20,7 +20,7 @@ class heap_card
         card_gen* pick_one();//fini
         void add_one( card_gen* nc);//a faire(Noah)
         card_gen *get_card_x(int n);
-        int get_size(void);//a faire(Noah)
+        size_t get_size(void);//a faire(Noah)
         virtual void render(sf::RenderWindow& window) = 0;
 };
